backend/requirements.txt
fastapi
uvicorn
pydantic
alembic
sqlalchemy
asyncpg
python-jose
passlib[bcrypt]
python-dotenv
email-validator
PyJWT